Add: TextArea and fix page URL
