+++ src/components/Button/Submit/index.tsx
@@ -1,8 +1,12 @@
 import React from "react";
 import { sBtnSubmit } from "./style";
 
-const ButtonSubmit = ({ text }) => {
-  return <button className={sBtnSubmit}>{text}</button>;
+const ButtonSubmit = (props) => {
+  return (
+    <button {...props} className={sBtnSubmit}>
+      {props.children}
+    </button>
+  );
 };
 
 export default ButtonSubmit;
