+++ src/components/Form/index.ts
@@ -1,5 +1,6 @@
 import FormInput from "./Input";
-import FormTextArea from "./TextArea";
+import FormTextArea from "./TextArea/FormTextArea";
 import FormSelect from "./Select";
+import TextArea from "./TextArea/TextArea";
 
-export { FormInput, FormTextArea, FormSelect };
+export { FormInput, FormTextArea, FormSelect, TextArea };
